move spotlight registration form file to admin/auth
