+++ frontend/javascripts/admin/auth/spotlight_registration_form.js
